GregTech generator rates, greatly improving text search speed.
It was taking a few seconds for each letter on the filter select screen. Profiling showed that it was mainly looking up burn rates (fixed by caching values for items without NBT), when F3+H was enabled, it spent a lot of time reparsing the format strings, and there was a bit of overhead to invoking TileEntityIronFurnace.getFuelValueFor with reflection.