Removed Liteloader, fixed IC², misc. warning corrections
Some part of IC² recipe support was still using Reflection (probably for API
version compatibility), so it wasn't an immediate error. Fixed after the
incompatibility showed in testing.

It seems forgegradle can't build liteloader and forge in the same project
anymore (or the way to do so was very obscure), so for now I'm just dropping
the code entirely. I tried making it a subproject, but excluding all the
rest of the mod that references forge would have been a lot of work.
Especially since recent API changes directly reference FluidStacks, so I'd
either need an annotation to omit specific functions when building
liteloader (Rust's #[cfg(feature="forge")] would be great, there), or leave
the liteloader version with a far less convenient API. Just picking out the
rest of forge-reliant code would be enough work, plus finding the specific
character sequence to make gradle do what I want, that it's easier to put
it off for now.