Fixed filter direction? (Old uncommited change)
Looks like I had assumed that Deque.push()/pop() operated on the end of the
list, rather than the start, so presumably the history limit code was
removing the most recent once full.

I probably made the change, then tested it by continuing to play a modpack,
but forgot to write a commit for the fix afterwards.